+++ examples/apps/todomvc-rt/src/store/index.js
@@ -5,6 +5,10 @@ import { middlewares } from "./middlewares"
 import { types } from "./types"
 import { createReactStore } from "@inglorious/react-store"
 
-export const store = createStore({ types, entities, middlewares })
+export const store = createStore({
+  types,
+  entities,
+  middlewares,
+})
 
 export const { Provider, useSelector, useNotify } = createReactStore(store)