+++ examples/apps/todomvc/src/store/index.js
@@ -1,11 +1,7 @@
 import { createStore } from "@inglorious/store"
-import { connectDevTools } from "@inglorious/store/client/dev-tools"
 
 import { entities } from "./entities"
+import { middlewares } from "./middlewares"
 import { types } from "./types"
 
-export const store = createStore({ types, entities })
-
-if (import.meta.env.DEV) {
-  connectDevTools(store)
-}
+export const store = createStore({ types, entities, middlewares })