+++ docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @inglorious/docs
 
+## 1.9.4
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @inglorious/engine@10.0.0
+  - @inglorious/renderer-2d@9.0.0
+  - @inglorious/renderer-react-dom@9.0.0
+
 ## 1.9.3
 
 ### Patch Changes