+++ examples/apps/todomvc-rt/CHANGELOG.md
@@ -1,14 +1,13 @@
 # todomvc-cs
 
-## 1.2.0
+## 1.0.0
 
-### Minor Changes
+### Major Changes
 
 - Add real-time app example
 
 ### Patch Changes
 
-- Turn devtools enhancer into a middleware
 - Updated dependencies
 - Updated dependencies
   - @inglorious/store@6.1.0