+++ examples/apps/todomvc/CHANGELOG.md
@@ -1,5 +1,18 @@
 # todomvc
 
+## 1.1.0
+
+### Minor Changes
+
+- Add real-time app example
+
+### Patch Changes
+
+- Turn devtools enhancer into a middleware
+- Updated dependencies
+- Updated dependencies
+  - @inglorious/store@6.1.0
+
 ## 1.0.2
 
 ### Patch Changes